updated the error message
