+++ src/languages/en.js
@@ -595,7 +595,7 @@ export default {
             pleaseEnterValidLogin: 'Please ensure the email or phone number is valid (e.g. +15005550006).',
             pleaseEnterUniqueLogin: 'That user is already a member of this workspace.',
             genericFailureMessage: 'An error occurred inviting the user to the workspace, please try again.',
-            systemUserError: 'Please ensure email or phone number is not a system User.',
+            systemUserError: ({email}) => `Sorry, you cannot invite ${email} to a workspace.`,
             welcomeNote: ({workspaceName}) => `You have been invited to the ${workspaceName} Workspace! Download the Expensify mobile App to start tracking your expenses.`,
         },
         editor: {
